Fix invoice number sequencing and enhance test coverage
This commit improves invoice number generation reliability and adds comprehensive test coverage for the numbering system. Key changes include:

• Add unique constraint on invoice_number and tenant
• Implement retries and gap detection in invoice number generation
• Fix edge cases with number sequence handling
• Handle overlapping and conflicting invoice numbers
• Improve prefix and padding handling

Key test additions:
• Sequential number generation and formatting
• Prefix handling (including Unicode and special chars)
• High initial values and padding behavior
• Number sequence integrity after failures
• Overlapping and conflicting numbers
• Edge cases for prefixes and number transitions

This significantly improves the robustness of invoice number generation while maintaining sequence integrity.
